Belajar Middleware - Membatasi Hak Akses User Multi Role

+++ resources/views/siswa.blade.php
@@ -12,7 +12,9 @@
 
         <div class="mb-3 d-flex justify-content-between">
             <a href="/siswa-add" class="btn btn-primary">Tambah Data <i class="fa-solid fa-plus"></i></a>
-            <a href="/siswa-deleted" class="btn btn-danger">Show Deleted Data</a>
+            @if (Auth::user()->role_id == 1)
+                <a href="/siswa-deleted" class="btn btn-danger">Show Deleted Data</a>    
+            @endif
         </div>
 
         @if (Session::has('status'))
@@ -51,38 +53,44 @@
                         <td>{{ $data->class->nama }}</td>
                         <td class="w-15">
                             <div class="container d-flex justify-content-around reset">
-                                <a href="/siswa-detail/{{ $data->id }}"><i class="fa-regular fa-eye fa-lg"></i></a>
-                                <a href="/siswa-edit/{{ $data->id }}"><i
-                                        class="fa-solid fa-pen-to-square fa-lg"></i></a>
-                                <a href="/siswa-delete/{{ $data->id }}" data-bs-toggle="modal"
-                                    data-bs-target="#siswa-delete-/siswa/{{ $data->id }}"><i
-                                        class="fa-solid fa-trash-can"></i></a>
-                                <!-- Button trigger modal -->
+                                @if (Auth::user()->role_id != 1 && Auth::user()->role_id != 2)
+                                    -
+                                @else
+                                    <a href="/siswa-detail/{{ $data->id }}"><i class="fa-regular fa-eye fa-lg"></i></a>    
+                                    <a href="/siswa-edit/{{ $data->id }}"><i class="fa-solid fa-pen-to-square fa-lg"></i></a>
+                                @endif
+                                
+                                @if (Auth::user()->role_id == 1)
+                                    <a href="/siswa-delete/{{ $data->id }}" data-bs-toggle="modal"
+                                        data-bs-target="#siswa-delete-/siswa/{{ $data->id }}"><i
+                                            class="fa-solid fa-trash-can"></i></a>
+                                    <!-- Button trigger modal -->
 
-                                <!-- Modal -->
-                                <div class="modal fade" id="siswa-delete-/siswa/{{ $data->id }}" tabindex="-1"
-                                    aria-labelledby="exampleModalLabel" aria-hidden="true">
-                                    <div class="modal-dialog">
-                                        <div class="modal-content">
-                                            <div class="modal-header">
-                                                <button type="button" class="btn-close" data-bs-dismiss="modal"
-                                                    aria-label="Close"></button>
-                                            </div>
-                                            <div class="modal-body">
-                                                Apakah anda yakin untuk menghapus data siswa : {{ $data->name }}.?
-                                            </div>
-                                            <div class="modal-footer">
+                                    <!-- Modal -->
+                                    <div class="modal fade" id="siswa-delete-/siswa/{{ $data->id }}" tabindex="-1"
+                                        aria-labelledby="exampleModalLabel" aria-hidden="true">
+                                        <div class="modal-dialog">
+                                            <div class="modal-content">
+                                                <div class="modal-header">
+                                                    <button type="button" class="btn-close" data-bs-dismiss="modal"
+                                                        aria-label="Close"></button>
+                                                </div>
+                                                <div class="modal-body">
+                                                    Apakah anda yakin untuk menghapus data siswa : {{ $data->name }}.?
+                                                </div>
+                                                <div class="modal-footer">
 
-                                                <form action="/siswa-destroy/{{ $data->id }}" method="POST">
-                                                    @csrf
-                                                    @method('delete')
-                                                    <button type="submit" class="btn btn-danger">Delete</button>
-                                                </form>
-                                                <a href="/siswa" class="btn btn-primary">Cancel</a>
+                                                    <form action="/siswa-destroy/{{ $data->id }}" method="POST">
+                                                        @csrf
+                                                        @method('delete')
+                                                        <button type="submit" class="btn btn-danger">Delete</button>
+                                                    </form>
+                                                    <a href="/siswa" class="btn btn-primary">Cancel</a>
+                                                </div>
                                             </div>
                                         </div>
                                     </div>
-                                </div>
+                                @endif
                             </div>
                         </td>
                     </tr>
